opentrack, gui: rename options dialog impl filename
