+++ gui/options-dialog.hpp
